Update static file serving and refactor novel routes
- Changed static file serving path in server.js to '/public' for better organization.
- Refactored novel routes in routes.ts to import from the novel directory directly.
- Enhanced the novel index file to define and export routes for chapters and AI rewrite settings.

These changes improve the structure of the application and enhance the clarity of route management.
