+++ src/novel/index.ts
@@ -1,3 +1,18 @@
-export * from './novel.controller'
-export * from './novel.service'
-export * from './novel.routes'
+import { Router } from 'express'
+import {
+	listChapters,
+	readChapter,
+	listNovels,
+	getAIRewriteSettings,
+	setAIRewriteSettings,
+} from './novel.controller'
+
+const router = Router()
+
+router.get('/novels', listNovels)
+router.get('/novels/:novelId/chapters', listChapters)
+router.get('/novels/:novelId/chapters/:volume/:chapter', readChapter)
+router.get('/settings/ai-rewrite', getAIRewriteSettings)
+router.post('/settings/ai-rewrite', setAIRewriteSettings)
+
+export default router
